refactor: reorganize middlewares and improve UI components
- Move listenerMiddleware to dedicated middlewares folder
- Add Google and Lock icons for authentication methods
- Improve error message handling
- Update table configurations and event form
- Refine WelcomeSlideContent styles
- Update toast helper for better middleware integration

Closes #24
