+++ src/store/middlewares/listenerMiddleware.ts
@@ -1,7 +1,7 @@
 import { createListenerMiddleware } from '@reduxjs/toolkit'
 
-import { baseApi } from '../services/baseApi'
-import { authApi } from '../services/authApi'
+import { baseApi } from '../../services/baseApi'
+import { authApi } from '../../services/authApi'
 
 export const listenerMiddleware = createListenerMiddleware()
