+++ src/types/ui/chip.ts
@@ -3,7 +3,7 @@ import { ColorProgressType } from './task'
 
 export interface ChipProps {
   label: string
-  role: string
+  role?: string
   color?: ColorProgressType
   variant?: VariantType
   className?: string
